New Release
now tables and definitions are handled more intelligently.

now list of records and numbers are expressions.
and now tables and definitions save expressions.
also added a new kind  of command:
<<<
Extraction
which takes  the form
A<<<(B|condition)
which means take all records from  B given condition and give it to A, B now has all the  records minus the records given to A